unify colored and normal waveform widgets
This commit unifies the colored variants of big and preview waveform
with the base classes for blue waveforms. This change is done to
seamlessly support displaying blue waveforms when no colored waveforms
are available. When handing data to waveform widgets, a boolean
parameter "colored" makes the widget render the data as colored if True.
Support for getting blue waveforms when a request for colored data fails
is not implemented yet.